final commit

+++ src/main/java/com/ust/pos/user/UserController.java
@@ -1,10 +1,15 @@
 package com.ust.pos.user;
 
-import com.ust.pos.dto.RoleDto;
 import com.ust.pos.dto.UserDto;
+import com.ust.pos.node.service.NodeService;
 import com.ust.pos.role.service.RoleService;
 import com.ust.pos.user.service.UserService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -17,10 +22,13 @@ public class UserController {
     private UserService userService;
     @Autowired
     private RoleService roleService;
+    @Autowired
+    private NodeService nodeService;
 
     @GetMapping("/list")
     public String home(Model model) {
         model.addAttribute("users", userService.findAll());
+        model.addAttribute("nodes", nodeService.getNodesForRoles());
         return "user/list";
     }
 
@@ -44,8 +52,17 @@ public class UserController {
     }
 
     @GetMapping("/delete")
-    public String delete(Model model, @RequestParam String username) {
-        userService.delete(username);
+    public String delete(@RequestParam String username, HttpServletRequest request, HttpServletResponse response) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if(auth!=null) {
+            String currentUsername = auth.getName();
+            userService.delete(username);
+
+            if (currentUsername.equals(username)) {
+                new SecurityContextLogoutHandler().logout(request, response, auth);
+                return "redirect:/login?deleted=true";
+            }
+        }
         return "redirect:/user/list";
     }
 }
